test: share simulation lotus-core fake across suites
