+++ tests/e2e/test_smoke.py
@@ -1,6 +1,7 @@
 from fastapi.testclient import TestClient
 from app.main import app
 from tests.support.app_runtime import override_app_runtime
+from tests.support.lotus_core_fakes import SimulationLotusCoreClient
 from tests.support.returns_series_payloads import (
     JAN_2026_PORTFOLIO_RETURNS,
     JAN_2026_RISK_FREE_RETURNS,
@@ -248,108 +249,6 @@ def test_e2e_concentration_stateless_payload() -> None:
     assert body["risk_proxy"]["hhi_proposed"] == 6250.0
 
 
-class _FakeLotusCoreClient:
-    async def create_simulation_session(
-        self,
-        *,
-        portfolio_id: str,
-        ttl_hours: int | None,
-        created_by: str | None,
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        return {
-            "session": {
-                "session_id": "SIM_E2E_0001",
-                "portfolio_id": portfolio_id,
-                "status": "ACTIVE",
-                "version": 1,
-                "created_by": created_by,
-                "created_at": "2026-02-27T10:30:00Z",
-                "expires_at": "2026-02-28T10:30:00Z",
-            }
-        }
-
-    async def add_simulation_changes(
-        self,
-        *,
-        session_id: str,
-        changes: list[dict[str, object]],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        assert session_id == "SIM_E2E_0001"
-        assert len(changes) == 1
-        return {"session_id": session_id, "version": 2, "changes": []}
-
-    async def get_core_snapshot(
-        self,
-        *,
-        portfolio_id: str,
-        request_payload: dict[str, object],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        if request_payload.get("snapshot_mode") == "BASELINE":
-            return {
-                "portfolio_id": portfolio_id,
-                "as_of_date": "2026-02-27",
-                "snapshot_mode": "BASELINE",
-                "valuation_context": {
-                    "portfolio_currency": "EUR",
-                    "reporting_currency": "USD",
-                    "position_basis": "market_value_base",
-                    "weight_basis": "total_market_value_base",
-                },
-                "sections": {
-                    "positions_baseline": [
-                        {"security_id": "SEC_A", "market_value_base": "80"},
-                        {"security_id": "SEC_B", "market_value_base": "20"},
-                    ]
-                },
-            }
-        return {
-            "portfolio_id": portfolio_id,
-            "as_of_date": "2026-02-27",
-            "snapshot_mode": "SIMULATION",
-            "valuation_context": {
-                "portfolio_currency": "EUR",
-                "reporting_currency": "USD",
-                "position_basis": "market_value_base",
-                "weight_basis": "total_market_value_base",
-            },
-            "simulation": {
-                "session_id": "SIM_E2E_0001",
-                "version": 2,
-                "baseline_as_of_date": "2026-02-27",
-            },
-            "sections": {
-                "positions_baseline": [
-                    {"security_id": "SEC_A", "market_value_base": "60"},
-                    {"security_id": "SEC_B", "market_value_base": "40"},
-                ],
-                "positions_projected": [
-                    {"security_id": "SEC_A", "market_value_base": "90"},
-                    {"security_id": "SEC_B", "market_value_base": "10"},
-                ],
-            },
-        }
-
-    async def get_instrument_enrichment(
-        self,
-        *,
-        security_ids: list[str],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        return {
-            "records": [
-                {
-                    "security_id": security_id,
-                    "issuer_id": f"ISSUER_{security_id}",
-                    "ultimate_parent_issuer_id": f"UPI_{security_id}",
-                }
-                for security_id in security_ids
-            ]
-        }
-
-
 class _FakeLotusPerformanceClient:
     async def get_returns_series(
         self,
@@ -367,7 +266,13 @@ class _FakeLotusPerformanceClient:
 
 
 def test_e2e_concentration_stateful_mode() -> None:
-    with override_app_runtime(lotus_core_client=_FakeLotusCoreClient()):
+    with override_app_runtime(
+        lotus_core_client=SimulationLotusCoreClient(
+            session_id="SIM_E2E_0001",
+            simulation_version=2,
+            include_ultimate_parent_issuer_id=True,
+        )
+    ):
         client = TestClient(app)
         response = client.post(
             "/analytics/risk/concentration",
@@ -387,7 +292,13 @@ def test_e2e_concentration_stateful_mode() -> None:
 
 
 def test_e2e_concentration_simulation_mode() -> None:
-    with override_app_runtime(lotus_core_client=_FakeLotusCoreClient()):
+    with override_app_runtime(
+        lotus_core_client=SimulationLotusCoreClient(
+            session_id="SIM_E2E_0001",
+            simulation_version=2,
+            include_ultimate_parent_issuer_id=True,
+        )
+    ):
         client = TestClient(app)
         response = client.post(
             "/analytics/risk/concentration",
@@ -412,7 +323,11 @@ def test_e2e_concentration_simulation_mode() -> None:
 def test_e2e_rolling_metrics_stateful_mode() -> None:
     with override_app_runtime(
         lotus_performance_client=_FakeLotusPerformanceClient(),
-        lotus_core_client=_FakeLotusCoreClient(),
+        lotus_core_client=SimulationLotusCoreClient(
+            session_id="SIM_E2E_0001",
+            simulation_version=2,
+            include_ultimate_parent_issuer_id=True,
+        ),
     ):
         client = TestClient(app)
         response = client.post(
