+++ tests/integration/test_health.py
@@ -2,6 +2,7 @@ from fastapi.testclient import TestClient
 
 from app.main import app
 from tests.support.app_runtime import override_app_runtime
+from tests.support.lotus_core_fakes import SimulationLotusCoreClient
 
 
 def test_health_endpoints() -> None:
@@ -193,106 +194,13 @@ def test_concentration_rejects_legacy_payload_shape() -> None:
     assert response.status_code == 422
 
 
-class _FakeLotusCoreClient:
-    async def create_simulation_session(
-        self,
-        *,
-        portfolio_id: str,
-        ttl_hours: int | None,
-        created_by: str | None,
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        return {
-            "session": {
-                "session_id": "SIM_0001",
-                "portfolio_id": portfolio_id,
-                "status": "ACTIVE",
-                "version": 1,
-                "created_by": created_by,
-                "created_at": "2026-02-27T10:30:00Z",
-                "expires_at": "2026-02-28T10:30:00Z",
-            }
-        }
-
-    async def add_simulation_changes(
-        self,
-        *,
-        session_id: str,
-        changes: list[dict[str, object]],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        assert session_id == "SIM_0001"
-        assert len(changes) == 1
-        return {"session_id": session_id, "version": 3, "changes": []}
-
-    async def get_core_snapshot(
-        self,
-        *,
-        portfolio_id: str,
-        request_payload: dict[str, object],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        if request_payload.get("snapshot_mode") == "BASELINE":
-            return {
-                "portfolio_id": portfolio_id,
-                "as_of_date": "2026-02-27",
-                "snapshot_mode": "BASELINE",
-                "valuation_context": {
-                    "portfolio_currency": "EUR",
-                    "reporting_currency": "USD",
-                    "position_basis": "market_value_base",
-                    "weight_basis": "total_market_value_base",
-                },
-                "sections": {
-                    "positions_baseline": [
-                        {"security_id": "SEC_A", "market_value_base": "80"},
-                        {"security_id": "SEC_B", "market_value_base": "20"},
-                    ]
-                },
-            }
-        return {
-            "portfolio_id": portfolio_id,
-            "as_of_date": "2026-02-27",
-            "snapshot_mode": "SIMULATION",
-            "valuation_context": {
-                "portfolio_currency": "EUR",
-                "reporting_currency": "USD",
-                "position_basis": "market_value_base",
-                "weight_basis": "total_market_value_base",
-            },
-            "simulation": {
-                "session_id": "SIM_0001",
-                "version": 3,
-                "baseline_as_of_date": "2026-02-27",
-            },
-            "sections": {
-                "positions_baseline": [
-                    {"security_id": "SEC_A", "market_value_base": "60"},
-                    {"security_id": "SEC_B", "market_value_base": "40"},
-                ],
-                "positions_projected": [
-                    {"security_id": "SEC_A", "market_value_base": "90"},
-                    {"security_id": "SEC_B", "market_value_base": "10"},
-                ],
-            },
-        }
-
-    async def get_instrument_enrichment(
-        self,
-        *,
-        security_ids: list[str],
-        correlation_id: str | None,
-    ) -> dict[str, object]:
-        return {
-            "records": [
-                {"security_id": security_id, "issuer_id": f"ISSUER_{security_id}"}
-                for security_id in security_ids
-            ]
-        }
-
-
 def test_concentration_stateful_mode_uses_lotus_core_snapshot() -> None:
-    with override_app_runtime(lotus_core_client=_FakeLotusCoreClient()):
+    with override_app_runtime(
+        lotus_core_client=SimulationLotusCoreClient(
+            session_id="SIM_0001",
+            simulation_version=3,
+        )
+    ):
         client = TestClient(app)
         response = client.post(
             "/analytics/risk/concentration",
@@ -312,7 +220,12 @@ def test_concentration_stateful_mode_uses_lotus_core_snapshot() -> None:
 
 
 def test_concentration_simulation_mode_reuses_or_creates_session_and_returns_metadata() -> None:
-    with override_app_runtime(lotus_core_client=_FakeLotusCoreClient()):
+    with override_app_runtime(
+        lotus_core_client=SimulationLotusCoreClient(
+            session_id="SIM_0001",
+            simulation_version=3,
+        )
+    ):
         client = TestClient(app)
         response = client.post(
             "/analytics/risk/concentration",
